+++ highlighters/main/test-data/arith1.zsh
@@ -28,19 +28,10 @@
 # vim: ft=zsh sw=2 ts=2 et
 # -------------------------------------------------------------------------------------------------
 
-BUFFER=$': $((ls); (ls))'
+BUFFER=$': $(( 6 * 9 ))'
 
 expected_region_highlight=(
   '1 1 builtin' # :
-  '3 15 default' # $((ls); (ls))
-  '3 15 command-substitution-unquoted "issue #704"' # $((ls); (ls))
-  '3 4 command-substitution-delimiter-unquoted "issue #704"' # $(
-  '5 5 reserved-word "issue #704"' # (
-  '6 7 command "issue #704"' # ls
-  '8 8 reserved-word "issue #704"' # )
-  '9 9 commandseparator "issue #704"' # ;
-  '11 11 reserved-word "issue #704"' # (
-  '12 13 command "issue #704"' # ls
-  '14 14 reserved-word "issue #704"' # )
-  '15 15 command-substitution-delimiter-unquoted "issue #704"' # )
+  '3 14 default' # $(( 6 * 9 ))
 )
+expected_mismatch="currently the actual highlighting has one superfluous group that highlights the asterisk is highlighted as 'globbing'"